ARIES-768 Make the isolated application context manager more resilient to failure. Also don't try to uninstall bundles that are no longer installed
git-svn-id: https://svn.apache.org/repos/asf/aries/trunk@1190285 13f79535-47bb-0310-9956-ffa450edef68

+++ application/application-runtime-isolated/src/main/java/org/apache/aries/application/runtime/isolated/impl/ApplicationContextImpl.java
@@ -139,21 +139,26 @@ public class ApplicationContextImpl implements AriesApplicationContext
       for (Iterator<Bundle> bundleIter = _bundles.iterator(); bundleIter.hasNext();) {
         Bundle bundleToRemove = bundleIter.next();
   
-        try {
-          // If Bundle is active, stop it first.
-          if (bundleToRemove.getState() == Bundle.ACTIVE) {
-            _bundleFrameworkManager.stopBundle(bundleToRemove);
+        if (bundleToRemove.getState() != Bundle.UNINSTALLED) {
+          try {
+            // If Bundle is active, stop it first.
+            if (bundleToRemove.getState() == Bundle.ACTIVE) {
+              _bundleFrameworkManager.stopBundle(bundleToRemove);
+            }
+          } catch (BundleException be) {
+            LOGGER.debug(LOG_EXCEPTION, be);
           }
   
-          // Delegate the uninstall to the bundleFrameworkManager
-          _bundleFrameworkManager.uninstallBundle(bundleToRemove);
-  
-        } catch (BundleException be) {
-          LOGGER.debug(LOG_EXCEPTION, be);
-          throw be;
+          try {
+            // Delegate the uninstall to the bundleFrameworkManager
+            _bundleFrameworkManager.uninstallBundle(bundleToRemove);
+    
+          } catch (BundleException be) {
+            LOGGER.debug(LOG_EXCEPTION, be);
+          }
         }
+        bundleIter.remove();
       }
-      _bundles.clear();
       
       _state = ApplicationState.UNINSTALLED;
     }
